Create local image for loading images from paths (#627)
